fix: Update test specs and fix TypeScript errors
- Update to-row.spec.ts to include multi_location and reason_code fields
- Fix tally-cards column tests to match actual columns (warehouse_name, updated_at_pretty)
- Update tally-cards filter tests to expect 2 filters (status and updated)
- Add sections support to FormConfig type
- Fix TypeScript errors: apiEndpoint, showMultiBadge, format types, reason_code type assertion
- Fix warehouse filter dropdown type annotations
- Update resource config field types (boolean->bool, smallint->int, date->timestamp)
- Fix edit-with-tabs spread type error
- Remove unused props from EditFormClient

+++ src/lib/data/resources/tcm_user_tally_card_entry_locations.config.ts
@@ -41,7 +41,7 @@ const tcm_user_tally_card_entry_locations: ResourceConfig<
       entry_id: { type: "uuid", write: true },
       location: { type: "text", write: true },
       qty: { type: "int", write: true },
-      pos: { type: "smallint", nullable: true, write: true },
+      pos: { type: "int", nullable: true, write: true },
     },
   },
 };
